Added state-level analysis in U.S. using regionmask
This is just a template for how one would do sub-country level analysis for other countries. The only caveat is that here it was a bit easier because regionmask had already processed the U.S. state shapefiles behind the scenes. For other countries, we would have to do this ourselves (shouldn't be too hard).